reduce number of packages
