+++ packages/boom-nfts/README.md
@@ -1,1 +1,3 @@
 ### Boom NFT mint contracts
+
+See [https://clarity-deployed-contracts.netlify.app/protocols/boom-nfts/](https://clarity-deployed-contracts.netlify.app/protocols/boom-nfts/) for more details about deployed versions.
