+++ packages/common/contracts/mock-bns.clar
