+++ packages/common/contracts/nft-trait.clar
