+++ packages/common/contracts/sip-9-nft-trait.clar
