--- packages/core-surf-charity-cycle-24/.gitignore
@@ -1,4 +1,0 @@
-
-settings/Mainnet.toml
-settings/Testnet.toml
-history.txt
