--- packages/core-surf-charity-cycle-24/contracts/setup.clar
@@ -1,9 +1,0 @@
-(print (contract-call? .boombox-admin
-  add-boombox
-  .boombox-24
-  u1
-  u1
-  u40000000
-  {version: 0x01, hashbytes: 0x1234}
-  tx-sender))
-
